+++ src/forze_mongo/adapters/__init__.py
@@ -1,5 +1,6 @@
+from .document import MongoDocumentAdapter
 from .txmanager import MongoTxManagerAdapter, MongoTxScopeKey
 
 # ----------------------- #
 
-__all__ = ["MongoTxManagerAdapter", "MongoTxScopeKey"]
+__all__ = ["MongoDocumentAdapter", "MongoTxManagerAdapter", "MongoTxScopeKey"]